refactor(semantic): split dataflow module
